Add damerauLevenshtein matching to pager schedule queries
It's often the case that people have transposed a character or two when
querying hubot for pager information. We can make that slightly easier
on them by employing a variant of the Levenshtein distance algorithm
that takes into account transpositions (in addition to other spelling
mistakes).

Then, if we don't get an exact match, we can at least return a list of
suggestions.